refactor(redis-cloud): unify structure with flat module organization
- Move all model definitions from models/ subdirectory into their corresponding handler files
- Remove models/ directory completely to match redis-enterprise's flat structure
- Update all imports from crate::models:: to direct module imports
- Update lib.rs to re-export types from handler modules instead of models
- Fix all test imports to use the new structure
- Complete struct definitions for CloudPeering, CloudMetrics, and related types
- Maintain backward compatibility by re-exporting all public types

This change unifies the structure between redis-cloud and redis-enterprise crates,
making the codebase more consistent and easier to navigate.
