+++ crates/redis-cloud/src/database.rs
@@ -34,12 +34,95 @@
 //! # }
 //! ```
 
-use crate::{
-    Result,
-    client::CloudClient,
-    models::{CloudDatabase, CreateDatabaseRequest, UpdateDatabaseRequest},
-};
+use crate::{Result, client::CloudClient};
+use serde::{Deserialize, Serialize};
 use serde_json::Value;
+
+/// Represents a Redis Cloud database instance
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct CloudDatabase {
+    pub db_id: u32,
+    pub name: String,
+    pub protocol: String,
+    pub provider: String,
+    pub region: String,
+    pub status: String,
+    pub memory_limit_in_gb: f64,
+    pub support_oss_cluster_api: bool,
+    pub data_persistence: Option<String>,
+    pub replication: bool,
+    pub throughput_measurement_by: Option<String>,
+    pub throughput_measurement_value: Option<u32>,
+    pub auto_failover: Option<bool>,
+    pub activated_on: Option<String>,
+    pub last_modified: Option<String>,
+    pub public_endpoint: Option<String>,
+    pub private_endpoint: Option<String>,
+    pub replica_of: Option<String>,
+    pub clustering: Option<Clustering>,
+    pub security: Option<DatabaseSecurity>,
+
+    #[serde(flatten)]
+    pub extra: Value,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct Clustering {
+    pub number_of_shards: u32,
+    pub regex_rules: Option<Vec<String>>,
+    pub hashing_policy: Option<String>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct DatabaseSecurity {
+    pub password: Option<String>,
+    pub ssl_client_authentication: Option<bool>,
+    pub source_ips: Option<Vec<String>>,
+}
+
+/// Request to create a database
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct CreateDatabaseRequest {
+    pub name: String,
+    pub memory_limit_in_gb: f64,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub support_oss_cluster_api: Option<bool>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub data_persistence: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub replication: Option<bool>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub throughput_measurement_by: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub throughput_measurement_value: Option<u32>,
+}
+
+/// Request to update a database
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct UpdateDatabaseRequest {
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub name: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub memory_limit_in_gb: Option<f64>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub data_persistence: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub replication: Option<bool>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub throughput_measurement_by: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub throughput_measurement_value: Option<u32>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub data_eviction: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub password: Option<String>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct ThroughputMeasurement {
+    pub by: String,
+    pub value: u32,
+}
 
 /// Handler for Cloud database operations
 ///
@@ -223,6 +306,8 @@ impl CloudDatabaseHandler {
             replication: None,
             data_eviction: None,
             password: None,
+            throughput_measurement_by: None,
+            throughput_measurement_value: None,
         };
         self.update(subscription_id, database_id, request).await
     }
