+++ crates/redis-cloud/src/types.rs
@@ -1,2 +1,19 @@
-// Re-export all models under a `types` module for parity with redis-enterprise
-pub use crate::models::*;
+// Re-export types from handler modules for parity with redis-enterprise
+pub use crate::account::*;
+pub use crate::acl::*;
+pub use crate::api_keys::*;
+pub use crate::backup::*;
+pub use crate::billing::*;
+pub use crate::cloud_accounts::*;
+pub use crate::crdb::*;
+pub use crate::database::*;
+pub use crate::logs::*;
+pub use crate::metrics::*;
+pub use crate::peering::*;
+pub use crate::private_service_connect::*;
+pub use crate::region::*;
+pub use crate::sso::*;
+pub use crate::subscription::*;
+pub use crate::tasks::*;
+pub use crate::transit_gateway::*;
+pub use crate::users::*;
